It's a good idea to not delete locally hosted podcasts after you're done reading them.
git-svn-id: http://svn.habariproject.org/habari-extras@1249 f4a90942-1786-4aa1-b092-57c31935ddc3

+++ plugins/podcast/trunk/mp3info.php
@@ -122,7 +122,9 @@ class MP3Info
 			$this->duration = 0;
 		}
 
-		unlink( $tmp );
+		if( ! $this->is_local_file( $this->file_name ) ) {
+			unlink( $tmp );
+		}
 		return TRUE;
 	}
